# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface UserPreferences {
  /** @nullable */
  budgetMin?: number | null;
  /** @nullable */
  budgetMax?: number | null;
  travelTypes?: string[];
  /** @nullable */
  climate?: string | null;
  /** @nullable */
  activityLevel?: string | null;
}

export interface User {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  displayName: string;
  /** @nullable */
  profileImageUrl?: string | null;
  preferences?: UserPreferences;
  createdAt: string;
}

export interface UpdateUserBody {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  preferences?: UserPreferences;
}

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  pending: "pending",
  swiping: "swiping",
  matched: "matched",
  planning: "planning",
} as const;

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  inviteCode: string;
  status: GroupStatus;
  memberCount: number;
  createdAt: string;
  createdByUserId: string;
}

export type GroupDetailStatus =
  (typeof GroupDetailStatus)[keyof typeof GroupDetailStatus];

export const GroupDetailStatus = {
  pending: "pending",
  swiping: "swiping",
  matched: "matched",
  planning: "planning",
} as const;

export type GroupMemberRole =
  (typeof GroupMemberRole)[keyof typeof GroupMemberRole];

export const GroupMemberRole = {
  creator: "creator",
  member: "member",
} as const;

export interface GroupMember {
  id: number;
  userId: string;
  groupId: number;
  role: GroupMemberRole;
  hasCompletedPreferences: boolean;
  swipeCount: number;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  joinedAt: string;
}

export type DestinationScoreConsensusLevel =
  (typeof DestinationScoreConsensusLevel)[keyof typeof DestinationScoreConsensusLevel];

export const DestinationScoreConsensusLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface DestinationScore {
  destinationId: number;
  destinationName: string;
  country: string;
  imageUrl: string;
  groupScore: number;
  matchPercentage: number;
  consensusLevel: DestinationScoreConsensusLevel;
  likeCount: number;
  dislikeCount: number;
  superlikeCount: number;
  totalVotes: number;
}

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  inviteCode: string;
  status: GroupDetailStatus;
  memberCount: number;
  createdAt: string;
  createdByUserId: string;
  members: GroupMember[];
  topDestinations: DestinationScore[];
}

export interface CreateGroupBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateGroupBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface JoinGroupBody {
  inviteCode: string;
}

export interface LeaveGroupResult {
  success: boolean;
}

export interface SubmitPreferencesBody {
  budgetMin: number;
  budgetMax: number;
  travelTypes: string[];
  climate: string;
  activityLevel: string;
  /** @nullable */
  availableDates?: string | null;
}

export interface Destination {
  id: number;
  name: string;
  country: string;
  description: string;
  imageUrl: string;
  tags: string[];
  /** 1=budget, 2=mid, 3=luxury */
  costLevel: number;
  climateType: string;
  activityLevel: string;
  travelTypes: string[];
  avgRating: number;
}

export interface RecordSwipeBody {
  groupId: number;
  destinationId: number;
  /** -1=dislike, 1=like, 2=superlike */
  value: number;
}

export interface Swipe {
  id: number;
  userId: string;
  groupId: number;
  destinationId: number;
  value: number;
  createdAt: string;
}

export interface SwipeResult {
  swipe: Swipe;
  updatedScore: DestinationScore;
}

export interface MatchingResults {
  groupId: number;
  totalMembers: number;
  membersCompleted: number;
  completionPercent: number;
  topDestinations: DestinationScore[];
  consensusSummary: string;
  hasEnoughData: boolean;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  swipe: "swipe",
  join: "join",
  preference: "preference",
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  userId: string;
  displayName: string;
  /** @nullable */
  destinationName?: string | null;
  action: string;
  createdAt: string;
}

export interface GroupStats {
  groupId: number;
  totalSwipes: number;
  membersSwipedCount: number;
  membersPreferencesCount: number;
  topMatch?: DestinationScore;
  recentActivity: ActivityItem[];
}

export interface DashboardSummary {
  userId: string;
  activeGroups: number;
  totalSwipes: number;
  recentGroups: Group[];
  pendingInvites: number;
}

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};